requirements.txt
hydra-core
mlflow
numpy
omegaconf
pandas
PyYAML
scikit-learn
torch
torchvision
